cover: move provider to components
